Move QueryResults and ExecutionStats
